# openspec/changes/dcf-ai-feedback-architecture-v1/openapi.yaml
openapi: 3.0.3
info:
  title: DCF AI Feedback Platform API
  version: 0.1.0
  description: |
    Contract baseline for event-triggered feedback lifecycle, AI classification, and supervisor triage.
    Demo mode assumptions: static logins for auth, local-first runtime, SQLite-backed persistence.
    Hugging Face config baseline (adapted from DCF-Intake demo):
    LLM_PROVIDER=huggingface, HF_MODEL=Qwen/Qwen2.5-7B-Instruct,
    HF_API_BASE=https://router.huggingface.co/v1, LLM_REQUEST_TIMEOUT_MS=600000,
    token alias support via HF_API_TOKEN or HUGGINGFACE_API_KEY.
servers:
  - url: https://api.dcf-feedback.local/v1
security:
  - demoAuthHeader: []
...
paths:
...
  /triggers/events:
    post:
      tags: [triggers]
      summary: Ingest iFamilyNet trigger event
      operationId: ingestTriggerEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerEventIn'
      responses:
        '202':
          description: Event accepted for orchestration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerEventAccepted'
        '400':
          description: Invalid event payload
  /public/surveys/{token}:
    get:
      tags: [public]
      summary: Load survey for token (no auth)
      operationId: getPublicSurvey
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Survey payload for fill UI
        '410':
          description: Token expired or consumed
  /public/surveys/{token}/submit:
    post:
      tags: [public]
      summary: Submit survey via token (no auth)
      operationId: submitPublicSurvey
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                aiOptOut:
                  type: boolean
      responses:
        '201':
          description: Submission accepted
  /public/portal/surveys:
    get:
      tags: [public]
      summary: List open surveys for partner portal role
      operationId: listPortalSurveys
      security: []
      parameters:
        - in: query
          name: userGroup
          required: true
          schema:
            $ref: '#/components/schemas/UserGroup'
      responses:
        '200':
          description: Portal survey links
...
  /surveys/{surveyInstanceId}/submit:
    post:
      tags: [submissions]
      summary: Submit survey response
      operationId: submitSurvey
      parameters:
        - $ref: '#/components/parameters/SurveyInstanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackSubmissionIn'
      responses:
        '201':
          description: Submission accepted and queued for AI processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackSubmissionAccepted'
        '409':
          description: Duplicate or already submitted
  /submissions/{submissionId}/ai-result:
    get:
      tags: [submissions]
      summary: Retrieve AI processing outcome
      operationId: getAiResult
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '200':
          description: AI result payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiInference'
        '404':
          description: Submission or result not found
...
components:
  securitySchemes:
    demoAuthHeader:
      type: apiKey
      in: header
      name: x-demo-user
      description: Static demo login identity key that maps to a predefined role profile.
  parameters:
    SurveyInstanceId:
      name: surveyInstanceId
      in: path
      required: true
      schema:
        type: string
    SubmissionId:
      name: submissionId
      in: path
      required: true
      schema:
        type: string
  schemas:
    UserGroup:
      type: string
      enum: [mandated_reporter, volunteer, attorney, foster_parent]
    TriggerEventIn:
      type: object
      required: [eventId, triggerType, userGroup, occurredAt]
      properties:
        eventId:
          type: string
        triggerType:
          type: string
          enum: [fifty_one_a, volunteer_event, legal_milestone, placement_milestone, case_closure]
        userGroup:
          $ref: '#/components/schemas/UserGroup'
        filingId:
          type: string
        caseId:
          type: string
        placementId:
          type: string
        milestoneId:
          type: string
        occurredAt:
          type: string
          format: date-time
    TriggerEventAccepted:
      type: object
      properties:
        eventId:
          type: string
        orchestrationStatus:
          type: string
          enum: [accepted, duplicate_ignored]
...
    FeedbackSubmissionIn:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          additionalProperties: true
        aiOptOut:
          type: boolean
          default: false
    FeedbackSubmissionAccepted:
      type: object
      properties:
        submissionId:
          type: string
        aiProcessingStatus:
          type: string
          enum: [queued, processing, completed]
        inferenceProvider:
          type: string
          enum: [huggingface]
    AiInference:
      type: object
      properties:
        submissionId:
          type: string
        topic:
          type: string
        sentimentScore:
          type: number
          minimum: 1
          maximum: 5
        urgency:
          type: string
          enum: [low, medium, high]
        privilegeTagged:
          type: boolean
        explainabilitySummary:
          type: string
        recommendedRoute:
          type: string
          enum: [closed_loop_ai, supervisor_queue, legal_policy_queue]
        provider:
          type: string
          enum: [huggingface]
        timeoutMs:
          type: integer
          example: 600000
        retryCount:
          type: integer
          example: 2
...